backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.22
langgraph==0.2.74
langchain==0.3.25
langchain-core==0.3.81
langchain-community==0.3.27
langchain-google-genai==2.1.4
langchain-groq==0.3.2
openai==1.35.5
google-generativeai==0.8.5
spacy==3.7.4
PyMuPDF==1.24.5
pytesseract==0.3.10
Pillow==12.1.1
tavily-python==0.3.3
python-dotenv==1.0.1
pydantic==2.7.4
aiofiles==23.2.1
httpx==0.27.0
numpy==1.26.4